refactor: Unify utilities
